+++ src/main/java/io/electrum/moneytransfer/api/AdminResource.java
@@ -26,6 +26,7 @@ import javax.ws.rs.core.UriInfo;
 import io.electrum.moneytransfer.model.ErrorDetail;
 import io.electrum.moneytransfer.model.IdType;
 import io.electrum.moneytransfer.model.MoneyTransferAdminMessage;
+import io.electrum.moneytransfer.model.MoneyTransferCustomerOrderHistory;
 import io.electrum.moneytransfer.model.MoneyTransferFeeQuote;
 import io.electrum.moneytransfer.model.MoneyTransferQuoteRequest;
 import io.electrum.vas.model.ExchangeRate;
@@ -97,6 +98,32 @@ public abstract class AdminResource {
          public static final String TO_CURRENCY = "toCurrency";
          public static final String RECEIVER_ID = "receiverId";
          public static final String SETTLEMENT_ENTITY_ID = "settlementEntity";
+      }
+   }
+
+   public class GetCustomerOrderHistory {
+      private GetCustomerOrderHistory() {
+      }
+
+      public static final String RELATIVE_PATH = "/customers/orders";
+      public static final String FULL_PATH = AdminResource.PATH + RELATIVE_PATH;
+      public static final String GET_CUSTOMER_ORDER_HISTORY = "getCustomerOrderHistory";
+      public static final int SUCCESS = 200;
+
+      public class QueryParameters {
+         private QueryParameters() {
+         }
+
+         public static final String CUSTOMER_PROFILE_ID = "customerProfileId";
+         public static final String RECEIVER_ID = "receiverId";
+         public static final String MERCHANT_ID = "merchantId";
+         public static final String ORIGINATOR_INST_ID = "originatorInstId";
+         public static final String FROM_DATE_TIME = "fromDateTime";
+         public static final String TO_DATE_TIME = "toDateTime";
+         public static final String STATUS = "status";
+         public static final String PAYMENT_TYPE = "paymentType";
+         public static final String PAYMENT_IDENTIFIER_NAME = "paymentIdentifierName";
+         public static final String PAYMENT_IDENTIFIER_VALUE = "paymentIdentifierValue";
       }
    }
 
@@ -314,4 +341,53 @@ public abstract class AdminResource {
             uriInfo,
             httpServletRequest);
    }
+
+   @GET
+   @Path(GetCustomerOrderHistory.RELATIVE_PATH)
+   @Produces({ "application/json" })
+   @ApiOperation(value = GetCustomerOrderHistory.GET_CUSTOMER_ORDER_HISTORY, notes = "Queries the details of a client's existing money transfer orders.", response = MoneyTransferCustomerOrderHistory.class, authorizations = {
+         @Authorization(value = "httpBasic") }, tags = {})
+   @ApiResponses(value = {
+         @ApiResponse(code = AdminResource.GetCustomerOrderHistory.SUCCESS, message = "OK", response = MoneyTransferCustomerOrderHistory.class),
+         @ApiResponse(code = 400, message = "Bad request", response = ErrorDetail.class),
+         @ApiResponse(code = 500, message = "Internal Server Error", response = ErrorDetail.class),
+         @ApiResponse(code = 501, message = "Not implemented", response = ErrorDetail.class),
+         @ApiResponse(code = 503, message = "Service Unavailable", response = ErrorDetail.class),
+         @ApiResponse(code = 504, message = "Gateway Timeout", response = ErrorDetail.class) })
+   public final void getCustomerOrderHistory(
+         @ApiParam(value = "Uniquely identifies customer's profile on the upstream entity's system.", required = true) @QueryParam(GetCustomerOrderHistory.QueryParameters.CUSTOMER_PROFILE_ID) @NotNull String customerProfileId,
+         @ApiParam(value = "Identifies the service provider to whom this request must be directed.", required = true) @QueryParam(GetCustomerOrderHistory.QueryParameters.RECEIVER_ID) @NotNull String receiverId,
+         @ApiParam(value = "The assigned merchant identifier. Also known as card acceptor ID.") @QueryParam(GetCustomerOrderHistory.QueryParameters.MERCHANT_ID) @Pattern(regexp = MERCHANT_ID_REGEX) String merchantId,
+         @ApiParam(value = "Identifies the institution from which the transaction originates. Value to be assigned by Electrum.") @QueryParam(GetCustomerOrderHistory.QueryParameters.ORIGINATOR_INST_ID) String originatorInstId,
+         @ApiParam(value = "The date from which to start searching for orders (inclusive).") @QueryParam(GetCustomerOrderHistory.QueryParameters.FROM_DATE_TIME) String fromDateTime,
+         @ApiParam(value = "The date from which to end searching for orders (exclusive).") @QueryParam(GetCustomerOrderHistory.QueryParameters.TO_DATE_TIME) String toDateTime,
+         @ApiParam(value = "Should be among the options of the MoneyTransferLookupResponse Statuses.") @QueryParam(GetCustomerOrderHistory.QueryParameters.STATUS) String status,
+         @ApiParam(value = "Payment type to filter orders by.") @QueryParam(GetCustomerOrderHistory.QueryParameters.PAYMENT_TYPE) String paymentType,
+         @ApiParam(value = "Payment identifier for the payment type indicated. Required if 'paymentType' is defined.") @QueryParam(GetCustomerOrderHistory.QueryParameters.PAYMENT_IDENTIFIER_NAME) String paymentIdentifierName,
+         @ApiParam(value = "Payment identifier value for the payment type indicated. Required if 'paymentType' is defined.") @QueryParam(GetCustomerOrderHistory.QueryParameters.PAYMENT_IDENTIFIER_VALUE) String paymentIdentifierValue,
+         @Context SecurityContext securityContext,
+         @Context Request request,
+         @Suspended AsyncResponse asyncResponse,
+         @Context HttpHeaders httpHeaders,
+         @Context UriInfo uriInfo,
+         @Context HttpServletRequest httpServletRequest) {
+      getResourceImplementation().getCustomerOrderHistory(
+            customerProfileId,
+            merchantId,
+            originatorInstId,
+            fromDateTime,
+            toDateTime,
+            status,
+            paymentType,
+            paymentIdentifierName,
+            paymentIdentifierValue,
+            receiverId,
+            securityContext,
+            request,
+            httpHeaders,
+            asyncResponse,
+            uriInfo,
+            httpServletRequest);
+   }
+
 }